Move communication layer implementation into separate library
To make it usable for other projects.
