+++ apis/c/node/Cargo.toml
@@ -16,4 +16,3 @@ flume = "0.10.14"
 [dependencies.dora-node-api]
 default-features = false
 path = "../../rust/node"
-features = ["zenoh"]
