+++ apis/rust/node/src/config.rs
@@ -1,3 +1,4 @@
+use communication_layer_pub_sub::zenoh::zenoh_config;
 use once_cell::sync::OnceCell;
 use serde::{Deserialize, Serialize};
 use std::{
